backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Nathan Guan <nathanguan2007@gmail.com>"]
readme = "../README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.2"
uvicorn = "^0.30.6"
langchain = "^0.2.15"
qdrant-client = "^1.11.1"
modal = "^0.64.71"
langchain-community = "^0.2.15"
openai = "^1.43.0"
bs4 = "^0.0.2"
tiktoken = "^0.7.0"
langchain-qdrant = "^0.1.3"
python-dotenv = "^1.0.1"
langchain-openai = "^0.1.23"
python-multipart = "^0.0.10"
pypdf = "^5.0.0"
boto3 = "^1.35.24"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
